Updated the Python Script to Include some of the Popular Consoles
Now it is able to Download Gamecube, Nintendo Wii, Wii U, PS1, PS2, PS3, OG Xbox and Xbox 360.

These are all the "Mainline" Consoles, i would of added portables like the DS but they are not on this website, could potentially add a DS Route for a different website, if it's a similar solution.